organize code into components
